+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.1.2] - Unreleased
 - Encode granular features with base64
+- Refactor result format to generic ISCC data model
 
 ## [0.1.1] - 2024-06-25
 - Handle text decoding errors gracefully